Add : minio 연결 세팅
