+++ src/main/resources/application.yml
@@ -10,7 +10,7 @@ spring:
     # 데이터베이스 로그인 비밀번호
     password: lossleaderBack12!
     # 데이터베이스 url
-    url: jdbc:mysql://127.0.0.1:3306/lossleader
+    url: jdbc:mysql://114.202.45.254:3306/lossleader
   jpa:
     hibernate:
       # create-drop : 서버 실행할때 drop 및 생성을 실행하고 서버 종료시 drop 실행
